remove homepage
