+++ opsml/app/routes/router.py
@@ -13,7 +13,6 @@ from opsml.app.routes import (
     data,
     files,
     healthcheck,
-    homepage,
     metrics,
     models,
     projects,
@@ -31,7 +30,6 @@ def build_router(dependencies: Optional[Sequence[Any]] = None) -> APIRouter:
     api_router.include_router(files.router, tags=["file"], prefix="/opsml", dependencies=dependencies)
     api_router.include_router(data.router, tags=["data"], prefix="/opsml", dependencies=dependencies)
     api_router.include_router(audit.router, tags=["audit"], prefix="/opsml", dependencies=dependencies)
-    api_router.include_router(homepage.router, tags=["homepage"], dependencies=dependencies)
     api_router.include_router(registry.router, tags=["registry"], prefix="/opsml", dependencies=dependencies)
     api_router.include_router(projects.router, tags=["project"], prefix="/opsml", dependencies=dependencies)
     api_router.include_router(metrics.router, tags=["metrics"], prefix="/opsml", dependencies=dependencies)
